OTP updated to be responsive
