+++ insight-hub-dashboard/src/app/components/auth/auth.component.css
@@ -412,11 +412,12 @@ form {
 }
 
 .gender-option input[type="radio"]:checked + span {
-  background-color: #4f59ff;
-  color: white;
-  border-color: #4f59ff;
-  box-shadow: 0 4px 12px rgba(79, 89, 255, 0.4);
+  background-color: #d6d9ff;
+  color: #2e38f2;
+  border-color: #b6bbff;
+  box-shadow: inset 0 0 6px rgba(79, 89, 255, 0.3), 0 2px 8px rgba(79, 89, 255, 0.2);
 }
+
 
 .gender-option span:hover {
   background-color: #eaeaea;
@@ -545,5 +546,96 @@ form {
         box-shadow: 0 0 0 2px rgba(0, 123, 255, 0.25);
       }
     }
+  }
+}
+
+
+@media (max-width: 768px) {
+  .panel {
+    position: static; /* changed from absolute */
+    width: 100%;
+    height: auto;
+    padding: 20px;
+    transform: none !important;
+  }
+
+  .signup-inner-container {
+    flex-direction: column;
+    height: auto;
+  }
+
+  .left-panel,
+  .right-panel {
+    position: static;
+    width: 100%;
+    height: auto;
+    transform: none !important;
+    z-index: auto;
+  }
+}
+
+@media (max-width: 768px) {
+  .left-panel img {
+    width: 120px;
+  }
+
+  .main-button {
+    width: 100%;
+  }
+
+  .input-group input {
+    width: 100%;
+  }
+
+  .gender-options {
+    flex-direction: row;
+    gap: 0.5rem;
+  }
+}
+
+
+@media (max-width: 768px) {
+  .signup-inner-container {
+    flex-direction: column;
+    height: auto;
+    width: 90%;
+    max-width: 500px;
+    position: relative;
+  }
+
+  .panel {
+    width: 100%;
+    height: auto;
+    position: relative;
+    transition: transform 0.6s ease-in-out, opacity 0.6s ease-in-out;
+    opacity: 1;
+    z-index: 1;
+  }
+
+  .panel.slide-up {
+    transform: translateY(-100%);
+    opacity: 0;
+    z-index: 0;
+    pointer-events: none;
+  }
+
+  .panel.slide-down {
+    transform: translateY(0);
+    opacity: 1;
+    z-index: 2;
+  }
+
+  .left-panel, .right-panel {
+    width: 100%;
+    padding: 30px 20px;
+    transform: none !important;
+  }
+
+  .main-button {
+    width: 100%;
+  }
+
+  .input-group input {
+    width: 100%;
   }
 }
