+++ insight-hub-dashboard/src/app/components/forgot-password/forgot-password.component.css
@@ -1,0 +1,193 @@
+.signup-wrapper {
+    height: 100vh;
+    background: linear-gradient(135deg, #2e305e, #203a43, #4755d6);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-family: 'Segoe UI', sans-serif;
+  }
+  
+  .signup-inner-container {
+    position: relative;
+    width: 100vw;
+    max-width: 1200px;
+    height: 80vh;
+    background-color: #fff;
+    border-radius: 20px;
+    box-shadow: 0 8px 30px rgba(0, 0, 0, 0.15);
+    overflow: visible;
+  }
+  
+  form {
+    display: flex;
+    flex-direction: column;
+    gap: 12px;
+  }
+  
+  .input-group {
+    position: relative;
+    z-index: 1;
+  }
+  
+  .input-group i {
+    position: absolute;
+    top: 50%;
+    left: 12px;
+    transform: translateY(-50%);
+    color: #888;
+    font-size: 13px;
+    pointer-events: none;
+  }
+  
+  .input-group input {
+    width: 100%;
+    padding: 8px 12px 8px 36px;
+    border: none;
+    border-radius: 20px;
+    background-color: #f5f5f5;
+    font-size: 13px;
+    color: #333;
+    box-shadow: inset 0 2px 4px rgba(0, 0, 0, 0.05);
+    transition: box-shadow 0.3s, background-color 0.3s;
+  }
+  
+  .input-group input:focus {
+    background-color: #fff;
+    box-shadow: 0 0 0 3px rgba(79, 89, 255, 0.2);
+    outline: none;
+  }
+  
+  .main-button {
+    width: 100%;
+    padding: 10px;
+    background-color: #24253b;
+    color: white;
+    font-weight: bold;
+    font-size: 14px;
+    border: none;
+    border-radius: 20px;
+    cursor: pointer;
+    transition: 0.3s;
+  }
+  
+  .main-button:hover {
+    background-color: #040404;
+  }
+  
+  .tooltip-error-left {
+    position: absolute;
+    padding: 6px 10px;
+    font-size: 0.65rem;
+    border-radius: 6px;
+    z-index: 9999;
+    max-width: 200px;
+    color: white;
+    background-color: #be4b49;
+    display: block;
+    white-space: normal;
+    word-wrap: break-word;
+  }
+  
+  .tooltip-error-left::before {
+    content: '';
+    position: absolute;
+    top: 50%;
+    left: 100%;
+    transform: translateY(-50%);
+    border-top: 6px solid transparent;
+    border-bottom: 6px solid transparent;
+    border-left: 6px solid #be4b49;
+  }
+  
+  .form-block {
+    height: 100px;
+    display: flex;
+    justify-content: space-between;
+    flex-direction: column;
+  }
+  
+  .error-alert {
+    position: relative;
+    display: flex;
+    align-items: center;
+    background: #ffe6e6;
+    border: 1px solid #ff4d4f;
+    padding: 0.75rem 1rem;
+    border-radius: 0.5rem;
+    margin-top: 1rem;
+  
+    .error-icon {
+      margin-right: 0.5rem;
+      font-size: 1.25rem;
+    }
+  
+    .error-text {
+      flex: 1;
+      margin: 0;
+      color: #a80000;
+      font-weight: 500;
+  
+      .retry-link {
+        color: #d0021b;
+        text-decoration: underline;
+        cursor: pointer;
+  
+        &:hover {
+          color: #a80000;
+        }
+      }
+    }
+  
+    .close-btn {
+      background: transparent;
+      border: none;
+      font-size: 1.25rem;
+      line-height: 1;
+      color: #a80000;
+      cursor: pointer;
+      padding: 0;
+  
+      &:hover {
+        color: #d0021b;
+      }
+    }
+  }
+  
+  .otp-form {
+    .otp-inputs {
+      display: flex;
+      justify-content: center;
+      gap: 0.5rem;
+      margin-bottom: 1rem;
+  
+      .otp-box {
+        width: 2.5rem;
+        height: 2.5rem;
+        text-align: center;
+        font-size: 1.25rem;
+        border: 1px solid #ccc;
+        border-radius: 0.25rem;
+        outline: none;
+        transition: border-color 0.2s;
+  
+        &:focus {
+          border-color: #007bff;
+          box-shadow: 0 0 0 2px rgba(0, 123, 255, 0.25);
+        }
+      }
+    }
+  }
+  
+
+  .forgot-password-link {
+    color: #ff4f4f; /* Bootstrap's blue, or use any hex you prefer */
+    cursor: pointer;
+    text-decoration: none;
+    transition: color 0.3s ease;
+  
+    &:hover {
+      color: #ea1c1c; /* Darker blue on hover */
+    }
+  }
+
+
