+++ insight-hub-dashboard/src/app/components/forgot-password/forgot-password.component.html
@@ -1,17 +1,77 @@
-<div class="forgot-password-container">
-    <h2>Reset Your Password</h2>
-  
-    <form [formGroup]="forgotForm" (ngSubmit)="onSubmit()">
-  
-      <input type="email" placeholder="Email" formControlName="email" />
-      <input type="password" placeholder="Old Password" formControlName="oldPassword" />
-      <input type="password" placeholder="New Password" formControlName="newPassword" />
-      <input type="password" placeholder="Confirm Password" formControlName="confirmPassword" />
-      
-      <button type="button" (click)="sendOtp()">Send OTP</button>
-      <input type="text" placeholder="Enter OTP" formControlName="otp" />
-  
-      <button type="submit">Reset Password</button>
-    </form>
+<!-- Step 1: Email Form -->
+<div *ngIf="step === 1">
+  <h3>Find Your Account</h3>
+
+  <form [formGroup]="forgotPasswordForm" (ngSubmit)="onEmailSubmit()">
+    <div *ngIf="!verifiedUser" class="form-block">
+      <div class="input-group">
+        <i class="fas fa-envelope"></i>
+        <input type="email" placeholder="Enter your email" formControlName="email" />
+        <div class="tooltip-error-left" *ngIf="forgotPasswordForm.get('email')?.invalid && forgotPasswordForm.get('email')?.touched">
+          <small *ngIf="forgotPasswordForm.get('email')?.errors?.['required']">Email is required.</small>
+          <small *ngIf="forgotPasswordForm.get('email')?.errors?.['email']">Enter a valid email.</small>
+        </div>
+      </div>
+
+      <button class="main-button" type="submit" [disabled]="forgotPasswordForm.invalid || loading">
+        Search
+      </button>
+    </div>
+
+    <p class="small-text">
+      <a (click)="backToSignInClick()" class="forgot-password-link">Back to sign</a>
+    </p>
+  </form>
+
+  <div *ngIf="verifiedUser" class="user-preview-card">
+    <p><strong>Name:</strong> {{ verifiedUser.firstName }} {{ verifiedUser.lastName }}</p>
+    <p><strong>Email:</strong> {{ verifiedUser.email }}</p>
+    <button class="main-button" (click)="step = 2">Continue</button>
   </div>
-      
+
+  <div *ngIf="emailNotFound" class="error-alert">
+    <span class="error-icon">⚠️</span>
+    <p class="error-text">
+      Email not found. Please <a (click)="retry()" class="retry-link">try again</a>.
+    </p>
+    <button class="close-btn" (click)="emailNotFound = false">&times;</button>
+  </div>
+</div>
+
+<!-- Step 2: OTP Form -->
+<div *ngIf="step === 2">
+  <h3>Enter Verification Code</h3>
+  <form [formGroup]="otpForm" (ngSubmit)="onOtpSubmit()" class="otp-form">
+    <div class="otp-inputs">
+      <input
+        *ngFor="let ctrl of otpControls; let i = index"
+        type="text"
+        maxlength="1"
+        [formControlName]="ctrl"
+        (keyup)="onOtpKeyUp($event, i)"
+        (paste)="onOtpPaste($event)"
+        class="otp-box"
+        autocomplete="one-time-code"
+        inputmode="numeric"
+      />
+    </div>
+    <div class="tooltip-error-left" *ngIf="otpForm.invalid && otpForm.touched">
+      <small *ngIf="otpForm.errors?.['minlength']">
+        Enter a valid 6-digit code.
+      </small>
+    </div>
+
+    <button class="main-button" type="submit" [disabled]="otpForm.invalid || loading">
+      Verify OTP
+    </button>
+
+    <p class="small-text">
+      Didn’t receive the code?
+      <a (click)="resendOtp()" class="forgot-password-link">Resend</a>
+    </p>
+  </form>
+
+  <p class="small-text">
+    <a (click)="backToSignInClick()" class="forgot-password-link">Back to sign</a>
+  </p>
+</div>
